Refactor CCMLC2014 into model physics numerics and solver headers
